refactor(cli): split tui.rs into model/events/render modules (#1130)
Behavior-preserving module split per epic #1116. tui.rs (2,423 lines)
becomes tui/mod.rs (entry point + re-export) plus three concern modules:

- tui/model.rs: app-state structs (PanelVisibility, FocusPanel,
  TablesBrowserState, ResultsTableState, LayoutAreas, TuiApp) and the
  state transitions on them (execute_query, navigate_history, load/scan
  tables, refresh_metrics) plus format_duration/extract_table_name. All
  43 unit tests relocate here next to the code they exercise.
- tui/events.rs: crossterm event loop (run_tui) and per-panel key handlers.
- tui/render.rs: ratatui layout + widget drawing (ui, render_*, draw_help).

Shared items use pub(super) so siblings see them; public surface
(start_tui_mode) is unchanged. No semantic changes.
